Updates mi_image_view.MIImageView with latest changes
- Deletes pg_image_view
- Updates the usages

Fixes #723
